feat: added Grating base class and method to control holograms
Added base class grating making sure certain attributes are not
forgotten.
Added methods to PlaneHolograting (which is the base class for hologratings) to compute the construction points for the hologram and show the VLS law. There also are now methods for setting the hologram with Horiba, PyOptix or SOLEMIO conventions as well as the associated methods to get the hologram parameters in these convention however the grating was defined.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Changes to be committed:
#        classes.py
#